backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
groq
sentence-transformers
numpy
pypdf
youtube-transcript-api
Pillow
pytesseract
pdf2image
